LayoutTests/js/getOwnPropertyDescriptor-window-attributes-expected.txt

Tests that Object.getOwnPropertyDescriptor() works correctly for Window properties

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


* Window.screen
PASS descriptor.get is an instance of Function
PASS descriptor.set is an instance of Function
PASS descriptor.enumerable is true
FAIL descriptor.configurable should be true. Was false.
PASS descriptor.get.call(nonWindowObject) threw exception TypeError: The DOMWindow.screen getter can only be used on instances of DOMWindow.
PASS descriptor.get.call(window) === window.screen is true
FAIL descriptor.get.call() === window.screen should be true. Threw exception TypeError: The DOMWindow.screen getter can only be used on instances of DOMWindow

* Window.navigator
PASS descriptor.get is an instance of Function
PASS descriptor.set is an instance of Function
PASS descriptor.enumerable is true
FAIL descriptor.configurable should be true. Was false.
PASS descriptor.get.call(nonWindowObject) threw exception TypeError: The DOMWindow.navigator getter can only be used on instances of DOMWindow.
PASS descriptor.get.call(window) === window.navigator is true
FAIL descriptor.get.call() === window.navigator should be true. Threw exception TypeError: The DOMWindow.navigator getter can only be used on instances of DOMWindow

* Window.frameElement
PASS descriptor.get is an instance of Function
PASS descriptor.set is undefined.
PASS descriptor.enumerable is true
FAIL descriptor.configurable should be true. Was false.
PASS descriptor.get.call(nonWindowObject) threw exception TypeError: The DOMWindow.frameElement getter can only be used on instances of DOMWindow.
PASS descriptor.get.call(window) === window.frameElement is true
FAIL descriptor.get.call() === window.frameElement should be true. Threw exception TypeError: The DOMWindow.frameElement getter can only be used on instances of DOMWindow

* Window.name
PASS descriptor.get is an instance of Function
PASS descriptor.set is an instance of Function
PASS descriptor.enumerable is true
FAIL descriptor.configurable should be true. Was false.
PASS descriptor.get.call(nonWindowObject) threw exception TypeError: The DOMWindow.name getter can only be used on instances of DOMWindow.
PASS descriptor.get.call(window) === window.name is true
FAIL descriptor.get.call() === window.name should be true. Threw exception TypeError: The DOMWindow.name getter can only be used on instances of DOMWindow
PASS successfullyParsed is true

TEST COMPLETE
